==> benchmark/05_accuracy_per_genome_benchmark.ipynb <==
### Evaluate the classifier performance across different genomes (See EukRep paper, Fig 2C)

import pandas as pd

testing_res_tiara_euk = pd.read_csv("/fs/ess/PAS0439/MING/cilates_fungi_classifier/testset_res_tiara_eukrep.csv")
testing_res_guteuk = pd.read_csv("/fs/ess/PAS0439/MING/cilates_fungi_classifier/testset_res_guteuk.csv")

seq_origin = pd.read_csv("/fs/ess/PAS0439/MING/cilates_fungi_classifier/testset_seq_origin.csv")

testing_merged_tiara_euk = pd.merge(testing_res_tiara_euk, seq_origin, on = "seq")
testing_merged_guteuk = pd.merge(testing_res_guteuk, seq_origin, on = "seq")

fungi_genomes = set(testing_merged_guteuk.query('category == "fungi"').genome)
protozoa_genomes = set(testing_merged_guteuk.query('category == "protozoa"').genome) 
sags_genomes =  set(testing_merged_guteuk.query('category == "SAGs"').genome)
prokaryotes_genomes = set(testing_merged_guteuk.query('category == "prokaryotes"').genome)

genome_origin_tiara = []
genome_origin_eukrep = []
genome_origin_guteuk = []
accuracy_list_tiara = []
accuracy_list_eukrep = []
accuracy_list_guteuk = []
genome_list_tiara = []
genome_list_eukrep = []
genome_list_guteuk = []

# Tiara 
for fungi in fungi_genomes:
    df = testing_merged_tiara_euk.query('genome_x == @fungi')
    correct = 0
    incorrect = 0
    for index, row in df.iterrows():
        pred = row['tiara_out']
        Y = row['Y']
        if pred == Y:
            correct += 1
        else:
            incorrect += 1
    
    accuracy = correct / (correct + incorrect)
    genome_origin_tiara.append("fungi")
    accuracy_list_tiara.append(accuracy)
    genome_list_tiara.append(fungi)

for protozoa in protozoa_genomes:
    df = testing_merged_tiara_euk.query('genome_x == @protozoa')
    correct = 0
    incorrect = 0
    for index, row in df.iterrows():
        pred = row['tiara_out']
        Y = row['Y']
        if pred == Y:
            correct += 1
        else:
            incorrect += 1
    
    accuracy = correct / (correct + incorrect)
    genome_origin_tiara.append("protozoa")
    accuracy_list_tiara.append(accuracy)
    genome_list_tiara.append(protozoa)

for sag in sags_genomes:
    df = testing_merged_tiara_euk.query('genome_x == @sag')
    correct = 0
    incorrect = 0
    for index, row in df.iterrows():
        pred = row['tiara_out']
        Y = row['Y']
        if pred == Y:
            correct += 1
        else:
            incorrect += 1
    
    accuracy = correct / (correct + incorrect)
    genome_origin_tiara.append("SAG")
    accuracy_list_tiara.append(accuracy)
    genome_list_tiara.append(sag)

for proka in prokaryotes_genomes:
    df = testing_merged_tiara_euk.query('genome_x == @proka')
    correct = 0
    incorrect = 0
    for index, row in df.iterrows():
        pred = row['tiara_out']
        Y = row['Y']
        if pred == Y:
            correct += 1
        else:
            incorrect += 1
    
    accuracy = correct / (correct + incorrect)
    genome_origin_tiara.append("prokaryotes")
    accuracy_list_tiara.append(accuracy)
    genome_list_tiara.append(proka)

        


# EukRep
for fungi in fungi_genomes:
    df = testing_merged_tiara_euk.query('genome_x == @fungi')
    correct = 0
    incorrect = 0
    for index, row in df.iterrows():
        pred = row['eukrep_out']
        Y = row['Y']
        if pred == Y:
            correct += 1
        else:
            incorrect += 1
    
    accuracy = correct / (correct + incorrect)
    genome_origin_eukrep.append("fungi")
    accuracy_list_eukrep.append(accuracy)
    genome_list_eukrep.append(fungi)

for protozoa in protozoa_genomes:
    df = testing_merged_tiara_euk.query('genome_x == @protozoa')
    correct = 0
    incorrect = 0
    for index, row in df.iterrows():
        pred = row['eukrep_out']
        Y = row['Y']
        if pred == Y:
            correct += 1
        else:
            incorrect += 1
    
    accuracy = correct / (correct + incorrect)
    genome_origin_eukrep.append("protozoa")
    accuracy_list_eukrep.append(accuracy)
    genome_list_eukrep.append(protozoa)

for sag in sags_genomes:
    df = testing_merged_tiara_euk.query('genome_x == @sag')
    correct = 0
    incorrect = 0
    for index, row in df.iterrows():
        pred = row['eukrep_out']
        Y = row['Y']
        if pred == Y:
            correct += 1
        else:
            incorrect += 1
    
    accuracy = correct / (correct + incorrect)
    genome_origin_eukrep.append("SAG")
    accuracy_list_eukrep.append(accuracy)
    genome_list_eukrep.append(sag)

for proka in prokaryotes_genomes:
    df = testing_merged_tiara_euk.query('genome_x == @proka')
    correct = 0
    incorrect = 0
    for index, row in df.iterrows():
        pred = row['eukrep_out']
        Y = row['Y']
        if pred == Y:
            correct += 1
        else:
            incorrect += 1
    
    accuracy = correct / (correct + incorrect)
    genome_origin_eukrep.append("prokaryotes")
    accuracy_list_eukrep.append(accuracy)
    genome_list_eukrep.append(proka)

        


# GutEuk
for fungi in fungi_genomes:
    df = testing_merged_guteuk.query('genome == @fungi')
    correct = 0
    incorrect = 0
    for index, row in df.iterrows():
        pred = row['ensembled_out']
        Y = row['Y']
        if pred == Y:
            correct += 1
        else:
            incorrect += 1
    
    accuracy = correct / (correct + incorrect)
    genome_origin_guteuk.append("fungi")
    accuracy_list_guteuk.append(accuracy)
    genome_list_guteuk.append(fungi)

for protozoa in protozoa_genomes:
    df = testing_merged_guteuk.query('genome == @protozoa')
    correct = 0
    incorrect = 0
    for index, row in df.iterrows():
        pred = row['ensembled_out']
        Y = row['Y']
        if pred == Y:
            correct += 1
        else:
            incorrect += 1
    
    accuracy = correct / (correct + incorrect)
    genome_origin_guteuk.append("protozoa")
    accuracy_list_guteuk.append(accuracy)
    genome_list_guteuk.append(protozoa)

for sag in sags_genomes:
    df = testing_merged_guteuk.query('genome == @sag')
    correct = 0
    incorrect = 0
    for index, row in df.iterrows():
        pred = row['ensembled_out']
        Y = row['Y']
        if pred == Y:
            correct += 1
        else:
            incorrect += 1
    
    accuracy = correct / (correct + incorrect)
    genome_origin_guteuk.append("SAG")
    accuracy_list_guteuk.append(accuracy)
    genome_list_guteuk.append(sag)

for proka in prokaryotes_genomes:
    df = testing_merged_guteuk.query('genome == @proka')
    correct = 0
    incorrect = 0
    for index, row in df.iterrows():
        pred = row['ensembled_out']
        Y = row['Y']
        if pred == Y:
            correct += 1
        else:
            incorrect += 1
    
    accuracy = correct / (correct + incorrect)
    genome_origin_guteuk.append("prokaryotes")
    accuracy_list_guteuk.append(accuracy)
    genome_list_guteuk.append(proka)

        


accuracy_per_genome_tiara = pd.DataFrame({"genome_origin": genome_origin_tiara, "accuracy": accuracy_list_tiara, "genome": genome_list_tiara})
accuracy_per_genome_tiara.to_csv("/fs/ess/PAS0439/MING/cilates_fungi_classifier/stats_visualization/accuracy_per_genome_tiara.csv", index = None)

accuracy_per_genome_eukrep = pd.DataFrame({"genome_origin": genome_origin_eukrep, "accuracy": accuracy_list_eukrep, "genome": genome_list_eukrep})
accuracy_per_genome_eukrep.to_csv("/fs/ess/PAS0439/MING/cilates_fungi_classifier/stats_visualization/accuracy_per_genome_eukrep.csv", index = None)

accuracy_per_genome_guteuk = pd.DataFrame({"genome_origin": genome_origin_guteuk, "accuracy": accuracy_list_guteuk, "genome": genome_list_guteuk})
accuracy_per_genome_guteuk.to_csv("/fs/ess/PAS0439/MING/cilates_fungi_classifier/stats_visualization/accuracy_per_genome_guteuk.csv", index = None)
