==> benchmark/06_chop_contig_greater_than_50000bp_for_benchmark.ipynb <==
## chop contig greater than 50000bp when benchmark and record seq origin

import os
from Bio import SeqIO
import glob
from Bio.Seq import Seq
from Bio.SeqRecord import SeqRecord
from Bio import SeqIO
import pandas as pd

os.chdir("/fs/scratch/PAS0439/Ming/databases/gut_eukaryotes_classifier/test/")

def chopped_seq(seq_len):
    
    # List of choices
    splited = [3000, 4000, 5000, 8000, 10000, 20000]
    number_chunks = seq_len//50000
    remainder = seq_len%50000
    output_seq_length = []
    for f in range(number_chunks):
        output_seq_length += splited
    if remainder > 3000:
        output_seq_length.append(remainder)
        
    return output_seq_length

        
        

seq_list = []
genome_list = []
genome_type = []

proka = glob.glob("proka/*")
with open("proka.fasta", "w") as outfile:
    for f in proka:
        records = SeqIO.parse(f, "fasta")
        for record in records:
            if len(record.seq) < 3000:
                pass
            elif len(record.seq) < 50000:
                seq_list.append(record.id)
                genome_list.append(f.split("/")[1])
                genome_type.append("prokaryotes")
                SeqIO.write(record, outfile, 'fasta')
            else:
                output_seq_length = chopped_seq(len(str(record.seq)))
                index = 0
                seq_length = 0
                record_seq = str(record.seq)
                for i in output_seq_length:
                    new_seq = Seq(record_seq[:i])
                    new_id = f"{record.id}_{index+1}"
                    new_record =  SeqRecord(new_seq, id=new_id)
                    SeqIO.write(new_record, outfile, "fasta")
                    seq_list.append(new_id)
                    genome_list.append(f.split("/")[1])
                    genome_type.append("prokaryotes")
                    record_seq = record_seq[i:]
                    index += 1
                      
                    

fungi = glob.glob("fungi/*")
with open("fungi.fasta", "w") as outfile:
    for f in fungi:
        records = SeqIO.parse(f, "fasta")
        for record in records:
            if len(record.seq) < 3000:
                pass
            elif len(record.seq) < 50000:
                seq_list.append(record.id)
                genome_list.append(f.split("/")[1])
                genome_type.append("fungi")
                SeqIO.write(record, outfile, 'fasta')
            else:
                output_seq_length = chopped_seq(len(str(record.seq)))
                index = 0
                seq_length = 0
                record_seq = str(record.seq)
                for i in output_seq_length:
                    new_seq = Seq(record_seq[:i])
                    new_id = f"{record.id}_{index+1}"
                    new_record =  SeqRecord(new_seq, id=new_id)
                    SeqIO.write(new_record, outfile, "fasta")
                    seq_list.append(new_id)
                    genome_list.append(f.split("/")[1])
                    genome_type.append("fungi")
                    record_seq = record_seq[i:]
                    index += 1
                    

protozoa = glob.glob("protozoa/*")
with open("protozoa.fasta", "w") as outfile:
    for f in protozoa:
        records = SeqIO.parse(f, "fasta")
        for record in records:
            if len(record.seq) < 3000:
                pass
            elif len(record.seq) < 50000:
                seq_list.append(record.id)
                genome_list.append(f.split("/")[1])
                genome_type.append("protozoa")
                SeqIO.write(record, outfile, 'fasta')
            else:
                output_seq_length = chopped_seq(len(str(record.seq)))
                index = 0
                seq_length = 0
                record_seq = str(record.seq)
                for i in output_seq_length:
                    new_seq = Seq(record_seq[:i])
                    new_id = f"{record.id}_{index+1}"
                    new_record =  SeqRecord(new_seq, id=new_id)
                    SeqIO.write(new_record, outfile, "fasta")
                    seq_list.append(new_id)
                    genome_list.append(f.split("/")[1])
                    genome_type.append("protozoa")
                    record_seq = record_seq[i:]
                    index += 1
                      
                    

with open("sags_test_chopped.fasta", "w") as outfile:

    records = SeqIO.parse("sags_test.fa", "fasta")
    for record in records:
        if len(record.seq) < 3000:
            pass
        elif len(record.seq) < 50000:
            seq_list.append(record.id)
            genome_list.append(str(record.id).split("_")[0])
            genome_type.append("protozoa")
            SeqIO.write(record, outfile, 'fasta')
        else:
            output_seq_length = chopped_seq(len(str(record.seq)))
            index = 0
            seq_length = 0
            record_seq = str(record.seq)
            for i in output_seq_length:
                new_seq = Seq(record_seq[:i])
                new_id = f"{record.id}_{index+1}"
                new_record =  SeqRecord(new_seq, id=new_id)
                SeqIO.write(new_record, outfile, "fasta")
                seq_list.append(new_id)
                genome_list.append(new_id.split("_")[0])
                genome_type.append("protozoa")
                record_seq = record_seq[i:]
                index += 1

# sags = SeqIO.parse("sags_test.fa", "fasta")
# for record in sags:
#     if len(record.seq) >= 3000:
#         seq_list.append(str(record.id))
#         genome = str(record.id).split("_")[0]
#         genome_list.append(genome)
#         genome_type.append("SAG")

df = pd.DataFrame.from_dict({"sequence":seq_list, "genome": genome_list, "genome_type": genome_type})
df.to_csv("/fs/ess/PAS0439/MING/cilates_fungi_classifier/testset_seq_origin_chopped.csv", index = None)
